Simplify purchase_consignment and clean code
Until now, the consignment box could be checked on the product while unchecked on the category.
This is error prone and misleading for the user.

The consignment field is now related on the product and set directly on the category.

The code was also cleaned to reduce file lengths and respect the step down principle.
