+++ purchase_consignment/models/product_template.py
@@ -4,13 +4,19 @@
 from odoo import api, fields, models
 
 
-class ProductCategory(models.Model):
+class ProductTemplate(models.Model):
 
-    _inherit = 'product.category'
+    _inherit = 'product.template'
 
     consignment = fields.Boolean(
+        related="categ_id.consignment",
+        store=True,
         help="If checked, only one supplier can be selected in the supplier prices list "
-        "for products of this category. "
+        "for this product. "
         "The supplier will automatically be set as owner of the stock "
         "on receipt orders."
     )
+
+    @api.constrains('consignment')
+    def _check_single_consigned_vendor_if_consigned(self):
+        self.mapped('product_variant_ids')._check_single_consigned_vendor_if_consigned()
