+++ purchase_consignment_inventory/models/__init__.py
@@ -2,6 +2,6 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import (
-    product,
+    product_product,
     stock_inventory
 )
